Rank bugs fixed final version
**********************Level0 - The rank cannot be shown at the conclusion of a level as the new level appears instantly after completing the previous one.
**********************Level0-Rank+Hint - This one works because no new level/scene is associated with it.
--------------------------------------
Due to the instantaneous transition from one level to the next, the player's rank is unable to be displayed at the conclusion of a level. This issue does not affect "Level0-Rank+Hint," as it does not have an associated new level or scene, allowing for the rank and hint to be displayed without interruption.